Avoiding copy in "astype"
PiperOrigin-RevId: 940546536

+++ jax/_src/lax/lax.py
@@ -1726,7 +1726,7 @@ def _convert_element_type(
     operand = literals.TypedNdArray(arr, aval=aval)
   elif (isinstance(operand, np.ndarray) and operand.dtype != dtypes.float0 and
         new_dtype != dtypes.float0):
-    arr = operand.astype(new_dtype)
+    arr = operand.astype(new_dtype, copy=False)
     aval = core.ShapedArray(arr.shape, arr.dtype, weak_type=weak_type)
     operand = literals.TypedNdArray(arr, aval=aval)
